pyproject.toml
[tool.poetry]
name = "multigit"
version = "0.1.0"
description = "TODO" # TODO
authors = ["Jonathan Wight <jwight@mac.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.7.0"}
toml = "^0.10.2"
gitpython = "^3.1.29"

[tool.poetry.scripts]
multigit = "multigit.main:app"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
